Got rid of the QuadTemplate
It doesn't make sense to enforce particular signatures of Term-template
classes, especially their constructores. All in all there is only one
method needed for interoperability - equals(). And for being able to run
the Dataset tests it's enough to request an implementation to provide a
method returning the desired Quad template (same way it's already done
for DataFactory and Dataset).

To make it able to distinguish in the Dataset context betwen
non-matching quads and comparisons making no sense (also by a static
code analysis) a separate QuadCompare interface has been added.